Fix deletion of a resource by a CoAP request
After a DELETE request handler returns some processing from
the coap engine is still done with the resource, therefore
the resource cannot be deleted directly by oc_ri_delete_resource.
Instead it must be scheduled for deletion by
oc_delayed_delete_resource.
However, a resource scheduled for deletion must not be used once
the processing is done. Therefore it is removed from the list of
application resources and added to the list of resources scheduled
to be deleted.
To check whether your resource is scheduled to be deleted you can
use oc_ri_is_app_resource_to_be_deleted.

+++ security/oc_ael.c
@@ -85,7 +85,7 @@ static inline size_t
 oc_sec_ael_max_space(size_t device)
 {
   oc_sec_ael_t *a = &ael[device];
-  size_t res = OC_SEC_AEL_MAX_SIZE;
+  size_t res = (size_t)OC_SEC_AEL_MAX_SIZE;
   switch (a->unit) {
   case OC_SEC_AEL_UNIT_BYTE:
     res = a->maxsize;
@@ -147,7 +147,7 @@ oc_sec_ael_default(size_t device)
   oc_sec_ael_t *a = &ael[device];
   a->categoryfilter = OC_SEC_AEL_CATEGORYFILTER_DEFAULT;
   a->priorityfilter = OC_SEC_AEL_PRIORITYFILTER_DEFAULT;
-  a->maxsize = OC_SEC_AEL_MAX_SIZE;
+  a->maxsize = (size_t)OC_SEC_AEL_MAX_SIZE;
   a->unit = OC_SEC_AEL_UNIT_DEFAULT;
   a->events_size = 0;
   oc_sec_dump_ael(device);
